modelos-ia/embeddings/requirements.txt
google-genai>=1.14.0
datasets>=3.5.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0
soundfile>=0.12.1
scipy>=1.11.0
